fixed lemmatize_text
